Update fatture payload

+++ src/main/java/BW_U5/EPIC_ENERGY_SERVICES/entities/Fattura.java
@@ -21,7 +21,7 @@ public class Fattura {
     private long id;
 
     private LocalDate data;
-    private int importo;
+    private double importo;
     private int numero;
 
   @ManyToOne
